requirements.txt
Pillow
ipython
